Move dependency related classes to Transactron
