+++ coreblocks/fu/priv.py
@@ -6,6 +6,7 @@ from typing import Sequence
 
 from transactron import *
 from transactron.lib import BasicFifo
+from transactron.utils import DependencyManager
 
 from coreblocks.params import *
 from coreblocks.params.keys import MretKey
