+++ coreblocks/params/__init__.py
@@ -5,5 +5,4 @@ from .layouts import *  # noqa: F401
 from .fu_params import *  # noqa: F401
 from .keys import *  # noqa: F401
 from .icache_params import *  # noqa: F401
-from .dependencies import *  # noqa: F401
 from .instr import *  # noqa: F401
