+++ coreblocks/stages/func_blocks_unifier.py
@@ -2,8 +2,8 @@ from collections.abc import Iterable
 
 from amaranth import *
 
-from coreblocks.params import GenParams, BlockComponentParams, DependencyManager
-from coreblocks.params.dependencies import UnifierKey
+from coreblocks.params import GenParams, BlockComponentParams
+from transactron.utils.dependencies import UnifierKey, DependencyManager
 from transactron import Method, TModule
 from transactron.lib import MethodProduct, Collector
 from coreblocks.utils.protocols import Unifier
