+++ transactron/utils/__init__.py
@@ -4,3 +4,5 @@ from .debug_signals import *  # noqa: F401
 from .assign import *  # noqa: F401
 from .amaranth_ext import *  # noqa: F401
 from .transactron_helpers import *  # noqa: F401
+from .dependencies import *  # noqa: F401
+from .depcache import *  # noqa: F401
